I removed the additional GetPaymentTypeController file.
