--- BangazonApi/Models/Employees.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace firstSprint.Models
-{
-    public class Employees
-    {
-
-    }
-}
