--- BangazonApi/Models/PaymentTypes.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace firstSprint.Models
-{
-    public class PaymentTypes
-    {
-        public int Id { get; set; }
-        public int TypeAccountNumber { get; set; }
-        public string Type { get; set; }
-        public string BillingAddress { get; set; }
-        public Customers Customer { get; set; }
-
-
-    }
-}
