+++ BangazonApi/appsettings.json
@@ -4,5 +4,8 @@
       "Default": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=DESKTOP-DF4KNSB\\SQLEXPRESS;Database=Bangazon;Trusted_Connection=True;"
+  }
 }
